Launch SuperSU from Settings
If SuperSU is installed, show an entry in Settings below
"Developer options", which launches SuperSU when tapped.
At boot time SuperSU will detect if com.android.settings
contains the resources from this patch, and if so it
will hide itself from the launcher. This was implemented
in v1.85, so we only show the entry if the installed
SuperSU version is at least that.

Change-Id: I2af95a1a4396101ff87448cd0ae20addf235af52
(cherry picked from commit acb4c091f3953e6add7fc477baa7b98ec81ec3ff)

Conflicts:
	res/values/cr_strings.xml
	src/com/android/settings/SettingsActivity.java